# Conflicts:
#	iroh-unixfs/src/builder.rs

+++ iroh-unixfs/src/hamt.rs
@@ -317,7 +317,7 @@ pub(crate) fn bits(hash: &[u8; 8], pos: u32, len: u32) -> u32 {
 
 fn log2(x: u32) -> u32 {
     assert!(x > 0);
-    u32::BITS as u32 - x.leading_zeros() - 1
+    u32::BITS - x.leading_zeros() - 1
 }
 
 #[cfg(test)]
